backend/delete-pdf/requirements.txt
psycopg2-binary>=2.9.0
boto3>=1.26.0
PyJWT>=2.8.0
cryptography>=41.0.0